+++ typespec-extension/src/code-model-builder.ts
@@ -68,6 +68,7 @@ import {
   shouldGenerateProtocol,
   isInternal,
   SdkClient,
+  isInclude,
 } from "@azure-tools/typespec-client-generator-core";
 import { fail } from "assert";
 import {
@@ -366,7 +367,7 @@ export class CodeModelBuilder {
     if (access) {
       return access === "internal";
     } else {
-      // fallback to "internal", it will be
+      // TODO: deprecate "internal"
       return isInternal(context, operation);
     }
   }
@@ -377,31 +378,39 @@ export class CodeModelBuilder {
       const models: (Model | Enum)[] = Array.from(client.service.models.values());
       Array.from(client.service.enums.values()).forEach((it) => models.push(it));
 
+      // lambda to mark model as public
+      const modelAsPublic = (model: Model | Enum) => {
+        // check it does not contain Union
+        const union = unionReferredByType(this.program, model, this.typeUnionRefCache);
+        if (union) {
+          const errorMsg = `Model '${getTypeName(
+            model,
+            this.typeNameOptions,
+          )}' cannot be set as access=public, as it refers Union '${getUnionName(union, this.typeNameOptions)}'`;
+          throw new Error(errorMsg);
+        }
+
+        const schema = this.processSchema(model, model.name);
+
+        this.trackSchemaUsage(schema, {
+          usage: [SchemaContext.Public],
+        });
+      };
+
       for (const model of models) {
         if (!processedModels.has(model)) {
           const access = getAccess(model);
           if (access === "public") {
-            // check it does not contain Union
-            const union = unionReferredByType(this.program, model, this.typeUnionRefCache);
-            if (union) {
-              const errorMsg = `Model '${getTypeName(
-                model,
-                this.typeNameOptions,
-              )}' cannot be set as access=public, as it refers Union '${getUnionName(union, this.typeNameOptions)}'`;
-              throw new Error(errorMsg);
-            }
-
-            const schema = this.processSchema(model, model.name);
-
-            this.trackSchemaUsage(schema, {
-              usage: [SchemaContext.Public],
-            });
+            modelAsPublic(model);
           } else if (access === "internal") {
             const schema = this.processSchema(model, model.name);
 
             this.trackSchemaUsage(schema, {
               usage: [SchemaContext.Internal],
             });
+          } else if (model.kind === "Model" && isInclude(this.sdkContext, model)) {
+            // TODO: deprecate "include"
+            modelAsPublic(model);
           }
 
           const usage = getUsage(model);
@@ -596,12 +605,10 @@ export class CodeModelBuilder {
       },
     });
 
-    // TODO (weidxu): temporary disable codeModelOperation.internalApi
-    // codeModelOperation.internalApi = this.isInternal(this.sdkContext, operation);
-    const internalApi = this.isInternal(this.sdkContext, operation);
+    codeModelOperation.internalApi = this.isInternal(this.sdkContext, operation);
 
     const convenienceApiName = this.getConvenienceApiName(operation);
-    let generateConvenienceApi: boolean = !!convenienceApiName && !internalApi; // at present, internalApi means not convenienceApi. this could change.
+    let generateConvenienceApi: boolean = !!convenienceApiName;
 
     let apiComment: string | undefined = undefined;
     if (generateConvenienceApi) {
@@ -1259,7 +1266,12 @@ export class CodeModelBuilder {
             );
           });
 
-          this.trackSchemaUsage(optionBagSchema, { usage: [SchemaContext.Input, SchemaContext.Public] });
+          this.trackSchemaUsage(optionBagSchema, { usage: [SchemaContext.Input] });
+          if (op.internalApi) {
+            this.trackSchemaUsage(optionBagSchema, { usage: [SchemaContext.Internal] });
+          } else if (op.convenienceApi) {
+            this.trackSchemaUsage(optionBagSchema, { usage: [SchemaContext.Public] });
+          }
 
           // option bag parameter
           const optionBagParameter = new Parameter(